Updated changes to search employee by firstname/lastname/pincode delete user by id sort user by date if birth/date of joining

+++ src/main/java/com/neosoft/dto/EmployeeDto.java
@@ -7,7 +7,7 @@ public class EmployeeDto {
 	
 	private @Getter @Setter String city;
 	
-	private @Getter @Setter long pincode;
+	private @Getter @Setter int pincode;
 	
 	private @Getter @Setter String department;
